// src/app/core/models/user.model.ts
export interface User {
  id?: number;
  firstName: string;
  lastName: string;
  email: string;
  password?: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface UserRegister {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
}